Move knownnodes module into network package
