+++ src/shutdown.py
@@ -8,8 +8,8 @@ import state
 from debug import logger
 from helper_sql import sqlQuery, sqlStoredProcedure
 from inventory import Inventory
-from knownnodes import saveKnownNodes
 from network import StoppableThread
+from network.knownnodes import saveKnownNodes
 from queues import (
     addressGeneratorQueue, objectProcessorQueue, UISignalQueue, workerQueue)
